+++ connection.js
@@ -12,7 +12,7 @@ const constants = require('haraka-constants')
 const net_utils = require('haraka-net-utils')
 const Notes = require('haraka-notes')
 const utils = require('haraka-utils')
-const { Address } = require('address-rfc2821')
+const { Address } = require('./address')
 const ResultStore = require('haraka-results')
 
 // Haraka libs
@@ -1039,7 +1039,7 @@ class Connection {
         this.lognotice(dmsg, {
             code: constants.translate(retval === constants.cont ? constants.ok : retval),
             msg: msg || '',
-            sender: this.transaction.mail_from.address(),
+            sender: this.transaction.mail_from.address,
         })
         switch (retval) {
             case constants.deny:
@@ -1089,7 +1089,7 @@ class Connection {
             this.lognotice(dmsg, {
                 code: constants.translate(retval === constants.cont ? constants.ok : retval),
                 msg: msg || '',
-                sender: this.transaction.mail_from.address(),
+                sender: this.transaction.mail_from.address,
             })
         }
         switch (retval) {